Fix #420 - grass on yavin had corrupted colours.
This commit fixes a problem with corrupted colours on quick sprites,
when running as a 64-bit binary. unsigned longs were used for 4-byte
colours. While this would work on 32-bit systems, an unsigned long has 8
bytes on 64-bit systems. Colour types have been changed to uint32_t to
ensure the correct size on all systems.

The same has also been done for the GL_State function, which previously
accepted a bitfield of states as an unsigned long. Now it requires state
to be a uint32_t.

+++ code/rd-vanilla/tr_WorldEffects.cpp
@@ -455,7 +455,7 @@ private:
 	struct SWeatherZone
 	{
 		static bool	mMarkedOutside;		
-		ulong*		mPointCache;			// malloc block ptr
+		unsigned int*		mPointCache;			// malloc block ptr
 		
 		int			miPointCacheByteSize;	// size of block
 		SVecRange	mExtents;
@@ -611,8 +611,8 @@ public:
 			Wz.mHeight		=  (int)(Wz.mSize.mMaxs[1] - Wz.mSize.mMins[1]);
 			Wz.mDepth		= ((int)(Wz.mSize.mMaxs[2] - Wz.mSize.mMins[2]) + 31) >> 5;
 			
-			Wz.miPointCacheByteSize = (Wz.mWidth * Wz.mHeight * Wz.mDepth) * sizeof(ulong);
-			Wz.mPointCache  = (ulong *)Z_Malloc( Wz.miPointCacheByteSize, TAG_POINTCACHE, qtrue );
+			Wz.miPointCacheByteSize = (Wz.mWidth * Wz.mHeight * Wz.mDepth) * sizeof(unsigned int);
+			Wz.mPointCache  = (unsigned int *)Z_Malloc( Wz.miPointCacheByteSize, TAG_POINTCACHE, qtrue );
 		}
 		else
 		{
@@ -720,8 +720,8 @@ public:
 			CVec3		Mins;
 			int			x, y, z, q, zbase;
 			bool		curPosOutside;
-			ulong		contents;
-			ulong		bit;
+			unsigned int		contents;
+			unsigned int		bit;
 
 
 			// Record The Extents Of The World Incase No Other Weather Zones Exist
